requirements.txt
networkx~=3.4.2
joblib==1.4.2
xxhash~=3.5.0
tqdm~=4.67.1
loguru~=0.7.3
luaparser~=3.3.0
setuptools==76.0.0
pyvis~=0.3.2
PySide6>=6.5.0
PySide6-Addons>=6.5.0
Pillow>=10.0.0
selenium>=4.15.0